+++ src/algorithms/games/mod.rs
@@ -9,4 +9,5 @@ pub mod erdos_renyi;
 pub mod forestfire;
 pub mod grg;
 pub mod growing_random;
+pub mod islands;
 pub mod tree;