+++ src/lib.rs
@@ -106,6 +106,7 @@ pub use crate::algorithms::games::erdos_renyi::{erdos_renyi_gnm, erdos_renyi_gnp
 pub use crate::algorithms::games::forestfire::forest_fire_game;
 pub use crate::algorithms::games::grg::{grg_game, grg_game_with_coords};
 pub use crate::algorithms::games::growing_random::growing_random_game;
+pub use crate::algorithms::games::islands::simple_interconnected_islands_game;
 pub use crate::algorithms::games::tree::tree_game_lerw;
 pub use crate::algorithms::io::edgelist::read_edgelist;
 pub use crate::algorithms::operators::complementer::complementer;